+++ code_igniter/application/models/m_queries.php
@@ -68,24 +68,26 @@ class M_queries extends MY_Model
      */
     public function create($data = null)
     {
-        if (stripos($data->sql, 'update ') !== false OR stripos($data->sql, 'update`') !== false) {
-            log_error('ERR-0045', 'm_queries::create', 'SQL cannot contain UPDATE clause');
-            return false;
-        }
-        if (stripos($data->sql, 'delete from ') !== false OR stripos($data->sql, 'delete from`') !== false) {
-            log_error('ERR-0045', 'm_queries::create', 'SQL cannot contain DELETE clause.');
-            return false;
-        }
-        if (stripos($data->sql, 'insert into ') !== false OR stripos($data->sql, 'insert into`') !== false) {
-            log_error('ERR-0045', 'm_queries::create', 'SQL cannot contain INSERT clause.');
-            return false;
-        }
-        if (stripos($data->sql, 'where @filter') === false OR stripos($data->sql, 'where @filter or') !== false) {
-            // We don't have the HIGHLY RECOMMENDED @filter in our SQL
-            // Ensure the user creating this query has the admin role
-            if ( ! in_array('admin', $this->user->roles)) {
-                log_error('ERR-0022', 'queries::create');
+        if ( ! empty($data->sql)) {
+            if (stripos($data->sql, 'update ') !== false OR stripos($data->sql, 'update`') !== false) {
+                log_error('ERR-0045', 'm_queries::create', 'SQL cannot contain UPDATE clause');
                 return false;
+            }
+            if (stripos($data->sql, 'delete from ') !== false OR stripos($data->sql, 'delete from`') !== false) {
+                log_error('ERR-0045', 'm_queries::create', 'SQL cannot contain DELETE clause.');
+                return false;
+            }
+            if (stripos($data->sql, 'insert into ') !== false OR stripos($data->sql, 'insert into`') !== false) {
+                log_error('ERR-0045', 'm_queries::create', 'SQL cannot contain INSERT clause.');
+                return false;
+            }
+            if (stripos($data->sql, 'where @filter') === false OR stripos($data->sql, 'where @filter or') !== false) {
+                // We don't have the HIGHLY RECOMMENDED @filter in our SQL
+                // Ensure the user creating this query has the admin role
+                if ( ! in_array('admin', $this->user->roles)) {
+                    log_error('ERR-0022', 'queries::create');
+                    return false;
+                }
             }
         }
         if ($id = $this->insert_collection('queries', $data)) {
